+++ src/components/dialog/dialog-buttons.jsx
@@ -3,45 +3,47 @@ import injectSheet from 'react-jss';
 import PropTypes from 'prop-types';
 
 /**
- * A component which adds the appropriate styles for the content inside a modal.
+ * A component which adds the appropriate styles for a button row inside a dialog.
  *
  * @param {Object} props - The props for the component.
  * @param {Object} props.classes - Classes provided by Jss.
  * @param {String} props.className - Additional className.
- * @param {JSX} props.children - The actual content for the modal.
+ * @param {JSX} props.children - The buttons for the row.
  * @returns {JSX} - Returns the jsx.
  */
-export function ModalContent({
+export function ModalButtons({
   classes,
   className,
   children,
   ...props
 }) {
   return (
-    <main
-      className={`${classes.content} ${className}`}
+    <div
+      className={`${classes.buttons} ${className}`}
       {...props}
     >
       {children}
-    </main>
+    </div>
   );
 }
 
-ModalContent.propTypes = {
+ModalButtons.propTypes = {
   classes: PropTypes.shape({ content: PropTypes.string.isRequired }).isRequired,
   children: PropTypes.node.isRequired,
   className: PropTypes.string,
 };
 
-ModalContent.defaultProps = { className: '' };
+ModalButtons.defaultProps = { className: '' };
 
-ModalContent.styles = ({ modal: theme }) => {
-  return {
-    content: {
-      composes: 'modal--content',
-      padding: theme.padding,
-    },
-  };
+ModalButtons.styles = {
+  buttons: {
+    composes: 'dialog--buttons',
+    padding: 8,
+    display: 'flex',
+    flexDirection: 'row',
+    justifyContent: 'flex-end',
+    alignItems: 'center',
+  },
 };
 
-export default injectSheet(ModalContent.styles)(ModalContent);
+export default injectSheet(ModalButtons.styles)(ModalButtons);
