+++ src/components/dialog/dialog-content.jsx
@@ -3,43 +3,45 @@ import injectSheet from 'react-jss';
 import PropTypes from 'prop-types';
 
 /**
- * A component which adds the appropriate styles for the header of a modal.
+ * A component which adds the appropriate styles for the content inside a dialog.
  *
  * @param {Object} props - The props for the component.
  * @param {Object} props.classes - Classes provided by Jss.
  * @param {String} props.className - Additional className.
- * @param {JSX} props.children - The content for the header.
+ * @param {JSX} props.children - The actual content for the dialog.
  * @returns {JSX} - Returns the jsx.
  */
-export function ModalHeader({
+export function ModalContent({
   classes,
   className,
   children,
   ...props
 }) {
   return (
-    <header
-      className={`${classes.header} ${className}`}
+    <main
+      className={`${classes.content} ${className}`}
       {...props}
     >
       {children}
-    </header>
+    </main>
   );
 }
 
-ModalHeader.propTypes = {
+ModalContent.propTypes = {
   classes: PropTypes.shape({ content: PropTypes.string.isRequired }).isRequired,
   children: PropTypes.node.isRequired,
   className: PropTypes.string,
 };
 
-ModalHeader.defaultProps = { className: '' };
+ModalContent.defaultProps = { className: '' };
 
-ModalHeader.styles = {
-  header: {
-    composes: 'modal--header',
-    marginBottom: 20,
-  },
+ModalContent.styles = ({ dialog: theme }) => {
+  return {
+    content: {
+      composes: 'dialog--content',
+      padding: theme.padding,
+    },
+  };
 };
 
-export default injectSheet(ModalHeader.styles)(ModalHeader);
+export default injectSheet(ModalContent.styles)(ModalContent);
