+++ src/components/dialog/dialog.jsx
@@ -3,46 +3,21 @@ import PropTypes from 'prop-types';
 import injectSheet from 'react-jss';
 import { position } from 'polished';
 
-import EventHandler from '../event-handler';
-
 /**
  * The markup for the ModalContainer.
  *
  * @param {Object} props - Props for the component.
  * @param {Object} props.classes - Classes provided by Jss.
- * @param {Boolean} props.hasBackdrop - Whether or not the modal should have a backdrop.
- * @param {Boolean} props.closeOnBackdropClick - Whether or not the modal should be clicked
- * when the user clicks outside of the modal.
- * @param {Function} props.onBackdropPress - The callback when the user clicks outside of the modal.
- * @param {JSX} props.children - The content for the modal.
+ * @param {JSX} props.children - The content for the dialog.
  * @returns {JSX} - Returns the JSX.
  */
 export function Modal({
   classes,
-  hasBackdrop,
-  closeOnBackdropClick,
-  onBackdropPress,
   children,
 }) {
-  let backdrop = null;
-
-  if (hasBackdrop) {
-    backdrop = closeOnBackdropClick ? (
-      <EventHandler
-        component="span"
-        className={classes.backdrop}
-        onPress={onBackdropPress}
-      />
-    ) : (
-      <span className={classes.backdrop} />
-    );
-  }
-
   return (
     <div className={classes.modal}>
       {children}
-
-      {backdrop}
     </div>
   );
 }
@@ -53,12 +28,9 @@ Modal.propTypes = {
     backdrop: PropTypes.string.isRequired,
   }).isRequired,
   children: PropTypes.node.isRequired,
-  hasBackdrop: PropTypes.bool.isRequired,
-  closeOnBackdropClick: PropTypes.bool.isRequired,
-  onBackdropPress: PropTypes.func.isRequired,
 };
 
-Modal.styles = ({ modal: theme }) => {
+Modal.styles = ({ dialog: theme }) => {
   return {
     modal: {
       position: 'relative',
